third_party/WebKit/LayoutTests/platform/win7/external/wpt/IndexedDB/name-scopes-expected.txt

This is a testharness.js-based test.
FAIL Non-unique index keys assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL Unique index keys assert_equals: Expected success event, but got error event instead expected "success" but got "error"
Harness: the test ran to completion.
